Rename core module to tools
