+++ crapssim/strategy/tools.py
